pyproject.toml
[project]
name = "last-fm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "aiohttp-client-cache[sqlite]>=0.14.3",
    "anyio>=4.12.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
